backend/requirements.txt
fastapi
uvicorn
mongoengine
pydantic
passlib[bcrypt]
python-jose
python-dotenv
